Revert ChessBoard back to display flex.
- This allows the board to be much more lean logic wise.
- The chessboard in theory could even be an image, (just like Chess.com) does.
However, if we do this, then we would have to create a separate way of handling
the click events. I thought about 2 different ways to do so.
The first one would be to track clicks on the image element itself, which would
require us to compute the clicked cell based on the click position. This seems
really error prone so I would definitely avoid that.
The second one would be to create a div layer on top of the image with individual
div cells. This is more viable but I'm still not a fan of having a separate
image and click handlers. They could quite easily misalign, leading to bugs.

For these reasons I prefer to handle both the drawing and the clicking together.