Set correct File Letter / Rank Number positioning (#5)
* Revert ChessBoard back to display flex.

- This allows the board to be much more lean logic wise.
- The chessboard in theory could even be an image, (just like Chess.com) does.
However, if we do this, then we would have to create a separate way of handling
the click events. I thought about 2 different ways to do so:

The first one would be to track clicks on the image element itself, which would
require us to use the click position to compute the corresponding square coordinate.
This seems really error prone so I would definitely avoid that.

The second one would be to create an overlayed div layer on top of the image with
div cells arranged in a grid fashion. This is more viable but I'm still not a fan of
separating the board display from the click handlers, as they could easily
get misaligned, leading to bugs.

For these reasons I prefer to handle both the drawing and the clicking together.

* Install sass and sass-loader

* Handle square color entirely by CSS

* Create named grid template areas

* Use named grid area for highlight position

* Add rudimentary square click log visualization